email_service/requirements.txt
Flask
Flask-RESTful
pytest
python-dotenv
flasgger